+++ PruebaTecnicaOnOff.Api/appsettings.json
@@ -8,7 +8,9 @@
   "AllowedHosts": "*",
   "ConnectionStrings": {
     "connectionStringEF": "Data Source=localhost;Initial Catalog=PruebaTecnicaOnOff;Integrated Security=True;Trust Server Certificate=true",
-    "adoConnectionString": "Data Source=localhost;Initial Catalog=PruebaTecnicaOnOff;Integrated Security=True;",
+    "adoConnectionString": "Data Source=localhost;Initial Catalog=PruebaTecnicaOnOff;Integrated Security=True;"
+  },
+  "ApiKey": {
     "apiKeyValue": "tuClaveApiKeyAquí"
   }
 }
