+++ PruebaTecnicaOnOff.Application/Servicios/PremioSorteoService.cs
@@ -1,6 +1,6 @@
 ﻿using PruebaTecnicaOnOff.Application.Servicios.IServicios;
 using PruebaTecnicaOnOff.Core.Entities;
-using PruebaTecnicaOnOff.Infrastructure.Repository.IRepository;
+using PruebaTecnicaOnOff.Infrastructure.Repositorio.IRepositorio;
 using System.Text.RegularExpressions;
 
 
@@ -20,15 +20,15 @@ namespace PruebaTecnicaOnOff.Application.Servicios
             _premioSorteoRepository.InsertarPremio(premio);
         }
 
-        public string InsertarNumeroAsignado(NumeroAsignado numeroAsignado) 
+        public async Task<string> InsertarNumeroAsignado(NumeroAsignado numeroAsignado) 
         {
-			string numeroAsignar = GenerarNumeroAleatorio(numeroAsignado);
+			string numeroAsignar = await GenerarNumeroAleatorio(numeroAsignado);
 			numeroAsignado.Numero = numeroAsignar;
 			_premioSorteoRepository.InsertarNumeroAsignado(numeroAsignado);
 			return numeroAsignar;
 		}
 
-        private string GenerarNumeroAleatorio(NumeroAsignado numeroAsignado)
+        private async Task<string> GenerarNumeroAleatorio(NumeroAsignado numeroAsignado)
         {
 			string numero;
 			Random random = new();
@@ -36,7 +36,7 @@ namespace PruebaTecnicaOnOff.Application.Servicios
 			{
 				numero = random.Next(1, 100000).ToString().PadLeft(LONGITUD_NUMERO, '0');
 				numeroAsignado.Numero = numero;
-			} while (MasDeTresNumerosIguales(numero.ToString()) && _premioSorteoRepository.ValidarNumeroAsignado(numeroAsignado));
+			} while (MasDeTresNumerosIguales(numero.ToString()) && await _premioSorteoRepository.ValidarNumeroAsignado(numeroAsignado));
 
 			return numero;
 		}
