+++ PruebaTecnicaOnOff.Core/Entities/PremioSorteo.cs
@@ -11,8 +11,8 @@ namespace PruebaTecnicaOnOff.Core.Entities
 	{
 		[JsonIgnore]
 		public Guid Id { get; set; }
-		public string Name { get; set; }
-		public string? Description { get; set; }
-		public int? Amount { get; set; }
+		public string Nombre { get; set; }
+		public string? Descripcion { get; set; }
+		public int? Valor { get; set; }
 	}
 }
